changed vol and rating to doubles

+++ server/TheKaleCartelWebApi/TheKaleCartelWebApi/DTO/Beer/KaleBeerDetailsDto.cs
@@ -3,25 +3,20 @@ using System.ComponentModel.DataAnnotations;
 
 namespace TheKaleCartelWebApi.DTO.Beer
 {
-    public class KaleBeerDetailsDto
-    {
-
+  public class KaleBeerDetailsDto
+  {
     public int KaleBeerId { get; set; }
 
-      [Required]
-      public string Name { get; set; }
+    [Required] public string Name { get; set; }
 
-      [Required]
-      public string Description { get; set; }
+    [Required] public string Description { get; set; }
 
-      public string PictureUrl { get; set; }
+    public string PictureUrl { get; set; }
 
-      [Required]
-      public int Rating { get; set; }
+    [Required] public double Rating { get; set; }
 
-      [Required]
-      public int VolPercentage { get; set; }
-      public DateTime CreationDate { get; set; }
-      public int KaleProfileId { get; set; }
+    [Required] public double VolPercentage { get; set; }
+    public DateTime CreationDate { get; set; }
+    public int KaleProfileId { get; set; }
   }
 }
