+++ server/TheKaleCartelWebApi/TheKaleCartelWebApi/DTO/Recipies/KaleRecipeDto.cs
@@ -3,18 +3,16 @@ using System.ComponentModel.DataAnnotations;
 
 namespace TheKaleCartelWebApi.DTO.Recipies
 {
-    public class KaleRecipeDto
-    {
-        public int KaleRecipeId { get; set; }
+  public class KaleRecipeDto
+  {
+    public int KaleRecipeId { get; set; }
 
-        [Required]
-        public string Name { get; set; }
+    [Required] public string Name { get; set; }
 
-        [Required]
-        public int Rating { get; set; }
+    [Required] public double Rating { get; set; }
 
-      public DateTime CreationDate { get; set; }
+    public DateTime CreationDate { get; set; }
 
     public int KaleProfileId { get; set; }
-    }
+  }
 }
