+++ server/TheKaleCartelWebApi/TheKaleCartelWebApi/Models/KaleBeer.cs
@@ -6,9 +6,9 @@ namespace TheKaleCartelWebApi.Models
     {
         public int KaleBeerId { get; set; }
         public string Name { get; set; }
-        public int Rating { get; set; }
+        public double Rating { get; set; }
         public string PictureUrl { get; set; }
-        public int VolPercentage { get; set; }
+        public double VolPercentage { get; set; }
         public DateTime CreationDate { get; set; }
         public string Description { get; set; }
